Users backoffice crud init

+++ app/src/Controller/BackofficeController.php
@@ -13,8 +13,10 @@ use App\Slot\ReservationService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
 
 class BackofficeController extends AbstractController
@@ -277,21 +279,50 @@ class BackofficeController extends AbstractController
         // Dobavljanje svih korisnika sa rolom ROLE_USER
         $users = $this->userRepository->findByRole('ROLE_USER');
 
-        return $this->render('backoffice/users.html.twig', [
+        return $this->render('backoffice/user/users.html.twig', [
             'users' => $users,
         ]);
     }
 
-    #[Route('/backoffice/users/{id}/edit', name: 'user_edit')]
-    public function edit(User $user, Request $request): Response
+    #[Route('/backoffice/users/edit/{id}', name: 'user_edit')]
+    public function editUser(int $id, Request $request, UserPasswordHasherInterface $passwordHasher, EntityManagerInterface $em): Response
     {
-        dd('UNDER CONSTRUCTION');
-        // logika za editovanje korisnika
+        $user = $this->userRepository->find($id);
+
+        if (!$user) {
+            throw $this->createNotFoundException('User not found');
+        }
+
+        if ($request->isMethod('POST')) {
+            $email = $request->request->get('email');
+            $password = $request->request->get('password');
+
+            $user->setEmail($email);
+
+            if (!empty($password)) {
+                $hashedPassword = $passwordHasher->hashPassword($user, $password);
+                $user->setPassword($hashedPassword);
+            }
+
+            $em->flush();
+
+            return $this->redirectToRoute('backoffice_users');
+        }
+
+        return $this->render('backoffice/user/edit_user.html.twig', [
+            'user' => $user,
+        ]);
     }
 
-    #[Route('/backoffice/users/{id}/delete', name: 'user_delete')]
-    public function delete(User $user, EntityManagerInterface $em): Response
+    #[Route('/backoffice/users/delete/{id}', name: 'user_delete', methods: ['POST'])]
+    public function deleteUser(int $id, EntityManagerInterface $em): RedirectResponse
     {
+        $user = $this->userRepository->find($id);
+
+        if (!$user) {
+            throw $this->createNotFoundException('User not found');
+        }
+
         $em->remove($user);
         $em->flush();
